+++ Assets/Scripts/MagicAVoxel/Scripts/Vox/VoxAsset.cs
@@ -1,9 +1,7 @@
-using System.Collections.Generic;
 using UnityEngine;
 
 [System.Serializable]
 public class VoxAsset : ScriptableObject
 {
     public byte[] rawData;
-    public VoxData data;
 }